First revision of the Vector Map Editor
It works, but there are quite a few things that
could be improved on.

The Vector and Layer loading code should be seperated
from the Map Editor so that it can be shared with
the main game.

A more sophsticated way to modify, or at least
delete and redraw shapes is needed too.